requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pillow>=11.0.0
numpy>=2.0.0
reportlab>=4.2.0
matplotlib>=3.9.0
opencv-python-headless>=4.10.0.84